refactor(app-tools): compile the docs bundler from TypeScript
copy-docs.mjs was hand-written JS with a hand-maintained .d.mts beside it —
the declaration existed only because the module was not TypeScript. It now
lives in src/bundleDocs.ts and comes out of the normal rslib build, so the
types are generated and the pair of loose files under scripts/ goes away.

The three consumers keep their contracts:

- the `modern-bundle-docs` bin becomes a two-line CommonJS shim that dynamic-
  imports the compiled ESM and passes the package root in — the compiled
  module cannot know that path portably across output formats, and the shim
  keeps the shebang without needing a loader on user machines
- the `./bundle-docs` export points at the compiled ESM with generated types
- the package's own build runs the shim after rslib, same as before

The former direct-invocation realpath check is gone with the file: the CLI
only runs when the shim calls it, so import-vs-execute no longer needs
detecting.

The root gate keeps calling the bin; scripts/ at the repo root stays plain
node — that is its convention (check-dependencies.js, skipCI.js,
sync-skills.mjs), unlike EdenX's script workspaces.
